% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIFFtoJPG.R
\encoding{UTF-8}
\name{TIFFtoJPG}
\alias{TIFFtoJPG}
\title{TIFFtoJPG}
\usage{
TIFFtoJPG(input_raster)
}
\arguments{
\item{input_raster}{The input raster format image needed to be conversed.}
}
\description{
TIFFtoJPG
}
\examples{
# TIFFtoJPG(tiff_img)
}
