% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIFFtoPNG.R
\encoding{UTF-8}
\name{TIFFtoPNG}
\alias{TIFFtoPNG}
\title{TIFFtoPNG}
\usage{
TIFFtoPNG(input_raster)
}
\arguments{
\item{input_raster}{The input raster format image needed to be conversed.}
}
\description{
TIFFtoPNG
}
\examples{
# TIFFtoPNG(tiff_img)
}
